# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nova Launch API",
    "version": "1.0.0",
    "description": "REST API for the Nova Launch Stellar token deployment platform.",
    "contact": {
      "name": "Nova Launch",
      "url": "https://github.com/Emmyt24/Nova-launch"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://api.nova-launch.app",
      "description": "Production"
    }
  ],
...
  "paths": {
    "/auth/challenge": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Request a nonce / challenge to sign with your Stellar wallet",
        "description": "Returns a one-time nonce that the client must sign with their Stellar private key. The nonce expires after a short TTL. Corresponds to GET /auth/nonce in the NestJS controller.",
        "operationId": "authChallenge",
        "parameters": [
          {
            "name": "publicKey",
            "in": "query",
            "required": true,
            "description": "Stellar wallet public key (G-address, 56 characters)",
            "schema": {
              "type": "string",
              "minLength": 56,
              "maxLength": 56,
              "pattern": "^G[A-Z2-7]{55}$",
              "example": "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nonce issued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NonceResponse"
                },
                "example": {
                  "nonce": "3f7e1b2a-9c4d-4e8f-a1b2-c3d4e5f60718",
                  "expiresAt": 1719100800,
                  "message": "Sign this message to authenticate: 3f7e1b2a-9c4d-4e8f-a1b2-c3d4e5f60718"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid publicKey query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": "publicKey is required",
                  "error": "Bad Request"
                }
              }
            }
          },
          "429": {
            "description": "Too many nonce requests — rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Authenticate with Stellar wallet signature (verify challenge)",
        "description": "Verifies the Ed25519 signature over the nonce obtained from POST /auth/challenge. On success, returns a short-lived access token and a longer-lived refresh token. Corresponds to POST /auth/login in the NestJS controller.",
        "operationId": "authVerify",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletAuthRequest"
              },
              "examples": {
                "typical": {
                  "summary": "Standard wallet login",
                  "value": {
                    "publicKey": "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN",
                    "signature": "aBcDeFgHiJkLmNoPqRsTuVwXyZ0123456789+/aBcDeFgHiJkLmNoPqRsTuVwXyZ012=",
                    "nonce": "3f7e1b2a-9c4d-4e8f-a1b2-c3d4e5f60718"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful — tokens issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "accessToken": "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJHQUFaST...",
                  "refreshToken": "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJHQUFaST...",
                  "expiresIn": 900,
                  "tokenType": "Bearer",
                  "walletAddress": "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body — missing or invalid fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": ["publicKey should not be empty", "signature should not be empty"],
                  "error": "Bad Request"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired signature / nonce",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Invalid signature or nonce",
                  "error": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Account suspended or wallet blocked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Account is not authorised to authenticate",
                  "error": "Forbidden"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh access token using a valid refresh token",
        "description": "Accepts a non-expired refresh token and issues a new access token and rotated refresh token. The old refresh token is revoked on success.",
        "operationId": "authRefresh",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              },
              "examples": {
                "typical": {
                  "summary": "Standard refresh",
                  "value": {
                    "refreshToken": "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJHQUFaST..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New token pair issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "accessToken": "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJHQUFaSS...",
                  "refreshToken": "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJHQUFaSS...",
                  "expiresIn": 900,
                  "tokenType": "Bearer",
                  "walletAddress": "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed refreshToken field",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": ["refreshToken should not be empty"],
                  "error": "Bad Request"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token is expired or has been revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Refresh token is invalid or expired",
                  "error": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Refresh token belongs to a suspended account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Account is not authorised",
                  "error": "Forbidden"
                }
              }
            }
          }
        }
      }
    },
    "/auth/session": {
      "delete": {
        "tags": ["Authentication"],
        "summary": "Revoke current session (logout)",
        "description": "Revokes the JWT identified by the `jti` claim in the Bearer token, effectively ending the current session. Corresponds to POST /auth/logout in the NestJS controller.",
        "operationId": "authDeleteSession",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Session revoked — no content returned"
          },
          "401": {
            "description": "No valid Bearer token provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Unauthorized",
                  "error": "Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Token does not have permission to revoke this session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Forbidden",
                  "error": "Forbidden"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /auth/verify or POST /auth/refresh"
      }
    },
    "schemas": {
      "WalletAuthRequest": {
        "type": "object",
        "required": ["publicKey", "signature", "nonce"],
        "properties": {
          "publicKey": {
            "type": "string",
            "minLength": 56,
            "maxLength": 56,
            "pattern": "^G[A-Z2-7]{55}$",
            "description": "Stellar wallet public key (G-address, 56 characters)",
            "example": "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN"
          },
          "signature": {
            "type": "string",
            "description": "Ed25519 signature over the nonce, base64-encoded",
            "example": "aBcDeFgHiJkLmNoPqRsTuVwXyZ0123456789+/aBcDeFgHiJkLmNoPqRsTuVwXyZ012="
          },
          "nonce": {
            "type": "string",
            "format": "uuid",
            "description": "The nonce value returned by POST /auth/challenge",
            "example": "3f7e1b2a-9c4d-4e8f-a1b2-c3d4e5f60718"
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "A valid, non-expired JWT refresh token previously issued by POST /auth/verify or POST /auth/refresh",
            "example": "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJHQUFaST..."
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": ["accessToken", "refreshToken", "expiresIn", "tokenType", "walletAddress"],
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Short-lived JWT access token (Bearer). Valid for 15 minutes.",
            "example": "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJHQUFaST..."
          },
          "refreshToken": {
            "type": "string",
            "description": "Longer-lived JWT refresh token. Valid for 7 days.",
            "example": "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJHQUFaST..."
          },
          "expiresIn": {
            "type": "integer",
            "description": "Seconds until the access token expires.",
            "example": 900
          },
          "tokenType": {
            "type": "string",
            "description": "Token type — always 'Bearer'.",
            "example": "Bearer"
          },
          "walletAddress": {
            "type": "string",
            "description": "Stellar wallet address (G-address) of the authenticated user.",
            "example": "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN"
          }
        }
      },
      "NonceResponse": {
        "type": "object",
        "required": ["nonce", "expiresAt", "message"],
        "properties": {
          "nonce": {
            "type": "string",
            "format": "uuid",
            "description": "One-time nonce (UUID v4) to be signed by the wallet.",
            "example": "3f7e1b2a-9c4d-4e8f-a1b2-c3d4e5f60718"
          },
          "expiresAt": {
            "type": "integer",
            "description": "Unix timestamp (seconds) at which the nonce expires.",
            "example": 1719100800
          },
          "message": {
            "type": "string",
            "description": "Full challenge message that the wallet must sign.",
            "example": "Sign this message to authenticate: 3f7e1b2a-9c4d-4e8f-a1b2-c3d4e5f60718"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
    },
...
  }
}